Moved most DOM manipulations to a contained file
