+++ src/pages/index.html
@@ -50,6 +50,7 @@
 							<input
 								type="text"
 								id="relation-name"
+								name="relation"
 								placeholder="Search..."
 							/>
