+++ src/supplement/drawing.ts
@@ -1,5 +1,5 @@
-import { context } from "../script.js";
 import { degreesToPixels, metresToDegrees } from "./conversions.js";
+import { context } from "./dom.js";
 import { Coordinate } from "./index.js";
 import { offset, totalMultiplier } from "./view.js";
